+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # SEOmatic Changelog
 
+## 3.3.45 - 2021.07.20
+### Fixed
+* Fixed an issue where `MedicalBusiness` JSON-LD type didn't inherit all of the properties from `MedicalOrganization` that it should
+* Handle per-entry type settings for the sitemap generation (https://github.com/nystudio107/craft-seomatic/issues/929)
+
 ## 3.3.44 - 2021.07.09
 ### Added
 * Added `truncateTitleTags` & `truncateDescriptionTags` settings to control whether the max title / description lengths should be enforced or not (on by default) (https://github.com/nystudio107/craft-seomatic/issues/922)